source/conflictmanager/requirements-lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-lock.txt requirements.in
#
blinker==1.8.2
    # via flask
boto3==1.34.113
    # via -r requirements.in
botocore==1.34.113
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
flask==3.0.3
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
idna==3.8
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
numpy==2.0.1
    # via -r requirements.in
packaging==24.1
    # via gunicorn
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==2.2.2
    # via
    #   botocore
    #   requests
werkzeug==3.0.4
    # via flask
networkx==3.4.2

